+++ CHANGELOG.md
@@ -18,6 +18,8 @@ and this project attempts to adhere to [Semantic Versioning](https://semver.org/
 
 ## [Unreleased]
 
+## [0.17.2]
+
 ### Security
 
 - Normalize paths in asset manifest to prevent leaking system-specific information like Python version and installation paths
@@ -460,7 +462,7 @@ and this project attempts to adhere to [Semantic Versioning](https://semver.org/
 
 - Josh Thomas <josh@joshthomas.dev> (maintainer)
 
-[unreleased]: https://github.com/joshuadavidthomas/django-bird/compare/v0.17.1...HEAD
+[unreleased]: https://github.com/joshuadavidthomas/django-bird/compare/v0.17.2...HEAD
 [0.1.0]: https://github.com/joshuadavidthomas/django-bird/releases/tag/v0.1.0
 [0.1.1]: https://github.com/joshuadavidthomas/django-bird/releases/tag/v0.1.1
 [0.2.0]: https://github.com/joshuadavidthomas/django-bird/releases/tag/v0.2.0
@@ -500,3 +502,4 @@ and this project attempts to adhere to [Semantic Versioning](https://semver.org/
 [0.16.2]: https://github.com/joshuadavidthomas/django-bird/releases/tag/v0.16.2
 [0.17.0]: https://github.com/joshuadavidthomas/django-bird/releases/tag/v0.17.0
 [0.17.1]: https://github.com/joshuadavidthomas/django-bird/releases/tag/v0.17.1
+[0.17.2]: https://github.com/joshuadavidthomas/django-bird/releases/tag/v0.17.2